Redact albom fell + fixes

+++ back/deleteAlbum.php
@@ -7,7 +7,9 @@ catch(PDOException $e){
     throw new PDOException($e->getMessage(), (int)$e->getCode());
 }
 if(isset($_POST['id']) && isset($_POST['filePath'])){
-    unlink($_POST['filePath']);
+    if(checkDublicate($_POST['filePath'], $pdo, true) == false){
+        unlink($_POST['filePath']);
+    }
     $stmt = $pdo->prepare("DELETE FROM albums WHERE id=?");
     $stmt->bindParam(1, $_POST['id'], PDO::PARAM_INT);
     $stmt->execute();
@@ -20,4 +22,10 @@ if(isset($_POST['id']) && isset($_POST['filePath'])){
         $out = json_encode(['result' => "Success"]);
     }
     echo $out;
+}
+function checkDublicate($path, $pdo){
+    $stmt = $pdo->prepare("SELECT id FROM albums WHERE file_path = ?");
+    $stmt->bindParam(1, $path, PDO::PARAM_STR, 255);
+    $stmt->execute();       
+    return $stmt->rowCount() > 1 ? true : false;
 }
